+++ src/main/java/com/mycharge/trainingmanagementplatform/service/Impl/StudentServiceImpl.java
@@ -10,6 +10,8 @@ import com.mycharge.trainingmanagementplatform.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 
 @Service
 public class StudentServiceImpl implements StudentService {
@@ -24,9 +26,13 @@ public class StudentServiceImpl implements StudentService {
     @Override
     public Result find(JSONObject jsonObject) {
         try{
-            Result res = Result.getResult(1);
-            res.put("data",studentMapper.find(jsonObject));
-            return res;
+            if(jsonObject.getInteger("stu_id") != null) {
+                Result res = Result.getResult(1);
+                res.put("data", studentMapper.find(jsonObject));
+                return res;
+            }else{
+                return Result.getResult(0);
+            }
         }catch (Exception e){
             e.printStackTrace();
             return Result.getResult(0);
@@ -34,14 +40,51 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
+    public Result findByTeacher(JSONObject jsonObject) {
+        try{
+            if(jsonObject.getInteger("tea_id") != null) {
+                Result res = Result.getResult(1);
+                res.put("data", studentMapper.findByTeacher(jsonObject));
+                return res;
+            }
+            else if(jsonObject.getInteger("stu_id") != null){
+                Result res = Result.getResult(1);
+                List<JSONObject> list = studentMapper.find(jsonObject);
+                List<JSONObject> list2 =studentMapper.findStudentScore(jsonObject);
+                if(list2.size() > 0) {
+                    float allStudentScore = 0;
+                    for (int i = 0; i < list2.size(); i++) {
+                        allStudentScore += list2.get(i).getFloat("stu_score");
+                    }
+                    list.get(0).put("stu_score", allStudentScore / list2.size());
+                }else{
+                    list.get(0).put("stu_score","目前没有老师打分");
+                }
+                res.put("data",list);
+                return res;
+            }else{
+                return Result.getResult(0);
+            }
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResult(0);
+        }
+
+    }
+
+    @Override
     public Result update(JSONObject jsonObject) {
         try{
             if(jsonObject.getInteger("stu_id") != null) {
-                Result res = Result.getResult(1);
-                res.put("data", studentMapper.update(jsonObject));
-                jsonObject.put("acc_id", jsonObject.get("stu_id"));
-                accountMapper.update(jsonObject);
-                return res;
+                if(!studentMapper.find(jsonObject).isEmpty()) {
+                    Result res = Result.getResult(1);
+                    res.put("data", studentMapper.update(jsonObject));
+                    jsonObject.put("acc_id", jsonObject.get("stu_id"));
+                    accountMapper.update(jsonObject);
+                    return res;
+                }else{
+                    return Result.getResult(0);
+                }
             }else{
                 return Result.getResult(0);
             }
@@ -61,15 +104,18 @@ public class StudentServiceImpl implements StudentService {
     public Result delete(JSONObject jsonObject) {
         try{
             if(jsonObject.getInteger("stu_id") != null){
-                Result res = Result.getResult(1);
-                res.put("data",studentMapper.delete(jsonObject));
-                jsonObject.put("acc_id",jsonObject.getInteger("stu_id"));
-                accountMapper.delete(jsonObject);
-                return res;
+                if(!studentMapper.find(jsonObject).isEmpty()) {
+                    Result res = Result.getResult(1);
+                    res.put("data", studentMapper.delete(jsonObject));
+                    jsonObject.put("acc_id", jsonObject.getInteger("stu_id"));
+                    accountMapper.delete(jsonObject);
+                    return res;
+                }else{
+                    return Result.getResult(0);
+                }
             }else{
                 return Result.getResult(0);
             }
-
         }catch (Exception e){
             e.printStackTrace();
             return Result.getResult(0);
@@ -82,6 +128,23 @@ public class StudentServiceImpl implements StudentService {
             Result res = Result.getResult(1);
             res.put("data",schoolMapper.findAll());
             return res;
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResult(0);
+        }
+    }
+
+    @Override
+    public Result scoreForStudent(JSONObject jsonObject) {
+        try{
+            if(jsonObject.getInteger("tea_id") != null && jsonObject.getInteger("stu_id") != null
+                    && jsonObject.getFloat("stu_score") != null) {
+                Result res = Result.getResult(1);
+                res.put("data", studentMapper.scoreForStudent(jsonObject));
+                return res;
+            }else{
+                return Result.getResult(0);
+            }
         }catch (Exception e){
             e.printStackTrace();
             return Result.getResult(0);
