+++ src/main/java/com/mycharge/trainingmanagementplatform/service/Impl/TeacherServiceImpl.java
@@ -24,7 +24,7 @@ public class TeacherServiceImpl implements TeacherService {
     @Autowired
     private CompanyMapper companyMapper;
 
-
+    //在注册服务中实现
     @Override
     public Result insert(JSONObject jsonObject) {
         return null;
@@ -33,23 +33,9 @@ public class TeacherServiceImpl implements TeacherService {
     @Override
     public Result find(JSONObject jsonObject) {
         try{
-            Result res = Result.getResult(1);
-            res.put("data",teacherMapper.find(jsonObject));
-            return res;
-        }catch (Exception e){
-            e.printStackTrace();
-            return Result.getResult(0);
-        }
-    }
-
-    @Override
-    public Result update(JSONObject jsonObject) {
-        try{
             if(jsonObject.getInteger("tea_id") != null) {
                 Result res = Result.getResult(1);
-                res.put("data", teacherMapper.update(jsonObject));
-                jsonObject.put("acc_id", jsonObject.get("tea_id"));
-                accountMapper.update(jsonObject);
+                res.put("data", teacherMapper.find(jsonObject));
                 return res;
             }else{
                 return Result.getResult(0);
@@ -61,16 +47,42 @@ public class TeacherServiceImpl implements TeacherService {
     }
 
     @Override
-    public Result delete(JSONObject jsonObject) {
-                try{
+    public Result update(JSONObject jsonObject) {
+        try{
+            if(jsonObject.getInteger("tea_id") != null) {
+                if(!teacherMapper.find(jsonObject).isEmpty()) {
                     Result res = Result.getResult(1);
-                    if(jsonObject.getInteger("tea_id") != null){
-                        res.put("data",teacherMapper.delete(jsonObject));
-                        jsonObject.put("acc_id",jsonObject.getInteger("tea_id"));
-                        accountMapper.delete(jsonObject);
-                        return res;
-                    }
-                    else{
+                    res.put("data", teacherMapper.update(jsonObject));
+                    jsonObject.put("acc_id", jsonObject.get("tea_id"));
+                    accountMapper.update(jsonObject);
+                    return res;
+                }else{
+                    return Result.getResult(0);
+                }
+            }else{
+                return Result.getResult(0);
+            }
+        }catch (Exception e){
+            e.printStackTrace();
+            return Result.getResult(0);
+        }
+    }
+
+    @Override
+    public Result delete(JSONObject jsonObject) {
+        try{
+            if(jsonObject.getInteger("tea_id") != null){
+                if(!teacherMapper.find(jsonObject).isEmpty()) {
+                    Result res = Result.getResult(1);
+                    res.put("data", teacherMapper.delete(jsonObject));
+                    jsonObject.put("acc_id", jsonObject.getInteger("tea_id"));
+                    accountMapper.delete(jsonObject);
+                    return res;
+                }else{
+                    return Result.getResult(0);
+                }
+            }
+            else{
                 return Result.getResult(0);
             }
         }catch (Exception e){
